Final File Structure Commented some lines of code in SignUp Controller Files as they were showing some error Please comment the code as you code
