+++ README.md
@@ -1,1 +1,4 @@
-# app-mathx
+#### File Structure
+- Set `home` directory as the source directory
+- `main.mainClass` is the Main class
+
