+++ home/main/res/view/homepage.fxml
@@ -1,29 +1,22 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.ScrollPane?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.ColumnConstraints?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.layout.Pane?>
-<?import javafx.scene.layout.RowConstraints?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
 <?import javafx.scene.text.Font?>
-
-<GridPane alignment="center" hgap="10" vgap="10" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="main.java.controllers.homepageController">
+<GridPane alignment="center" hgap="10" vgap="10" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="main.src.controllers.homepageController">
    <columnConstraints>
       <ColumnConstraints />
    </columnConstraints>
    <rowConstraints>
       <RowConstraints />
    </rowConstraints>
-   <children>
+   
       <Pane prefHeight="720.0" prefWidth="1280.0">
          <opaqueInsets>
             <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
          </opaqueInsets>
-      <children>
+      
       <Label layoutX="1083.0" layoutY="14.0" prefHeight="43.0" prefWidth="183.0" text="Ria Mehta">
 
       </Label>
@@ -36,31 +29,27 @@
                   <Font name="Helvetica Bold" size="43.0" />
                </font></Label>
             <ScrollPane layoutX="51.0" layoutY="150.0" prefHeight="453.0" prefWidth="260.0">
-              <content>
                 <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="336.0" prefWidth="200.0">
-                     <children>
+                     
                         <Label layoutX="7.0" prefHeight="51.0" prefWidth="152.0" text="Submitted Assignments">
                            <font>
                               <Font name="Helvetica Bold" size="13.0" />
                            </font>
                         </Label>
-                     </children></AnchorPane>
-              </content>
+                     </AnchorPane>
             </ScrollPane>
             <ScrollPane layoutX="354.0" layoutY="150.0" prefHeight="453.0" prefWidth="260.0">
-               <content>
                   <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="336.0" prefWidth="200.0">
-                     <children>
+                     
                         <Label layoutX="7.0" prefHeight="51.0" prefWidth="158.0" text="Upcoming Assignments">
                            <font>
                               <Font name="Helvetica Bold" size="13.0" />
                            </font>
                         </Label>
-                     </children>
+                     
                   </AnchorPane>
-               </content>
             </ScrollPane>
-      </children>
+      
       </Pane>
-   </children>
+   
 </GridPane>
