+++ home/main/src/controllers/homepageController.java
@@ -1,4 +1,4 @@
-package main.java.controllers;
+package main.src.controllers;
 
 public class homepageController {
 }
